Renamed `set` to `vm_set`. Created `block_gas_limit_set` similar to `vm_set` folder. Removed `#[path = ...]`.
