+++ benches/benches/block_target_gas_set/alu.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::*;
 
 pub fn run_alu(group: &mut BenchmarkGroup<WallTime>) {
     run(
